Converting to clean arch
