+++ tarot/Data/Services/Service.swift
@@ -5,19 +5,7 @@
 //  Created by Fernando Marins on 18/09/24.
 //
 
-import Combine
 
-protocol ServiceProtocol {
-    func getCards() async throws -> [CardModel]
-    func getRunes() async throws -> [RuneModel]
-    func getDaemons() async throws -> [DaemonModel]
-    func getSangoma() async throws -> SangomaModel
-    func getAlphabet() async throws -> [LetterModel]
-    func getAstrology() async throws -> AstrologyModel
-    func getHerbs() async throws -> Herbs
-    func getHoodoo() async throws -> HoodooModel
-    func postHerb(_ herb: Herb) async throws -> Herb
-}
 
 class Service: ServiceProtocol {
     private let apiClient = URLSessionAPIClient<Endpoint>()
