+++ tarot/Domain/Models/AstrologyModel.swift
