+++ tarot/Domain/Models/CardModel.swift
