+++ tarot/Domain/Models/HoodooModel.swift
