+++ tarot/Domain/Models/RuneModel.swift
