+++ tarot/Domain/Models/SangomaModel.swift
